requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
cryptography==41.0.7
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10
langgraph==0.0.20
openai>=1.12.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
PyMySQL==1.1.0

